requirements.txt
azure-functions
azure-cosmos
azure-identity